+++ Src/Obfy.UI/DependencyInjection/AppModule.cs
@@ -4,7 +4,6 @@ using Obfy.Core.DependencyInjection;
 using Obfy.UI.Services;
 using Obfy.UI.ViewModels;
 using Obfy.UI.Views;
-using Obfy.UI.Views.Dialogs;
 
 namespace Obfy.UI.DependencyInjection;
 
@@ -65,9 +64,5 @@ public class AppModule : Module
         builder.RegisterType<MainWindow>()
             .AsSelf()
             .SingleInstance();
-
-        builder.RegisterType<AboutWindow>()
-            .AsSelf()
-            .InstancePerDependency();
     }
 }